feat: remove redundant layouts
